feat: move badge verification to pkg
